Don't force UTC on naive datetime structs
Why:
When a camera ships EXIF information that lacks any type of
offset and/or timezone information, any kind of information that
would help make a decision regarding the timezone is absent.
Forcing UTC means that most users of such cameras will end up
with datetimes reported that have nothing to do with the ones
they expect. The only assumption that holds some water, is to
assume that the picture was taken in the same timezone as the
device running Loupe and hope that's true. If it is not, it is
undetectable programmatically, and the user will have to mentally
compensate for it

What:
Switch from using from_unix_local() and parsing the timestamp of
the naive datetime as UTC to using from_local() and passing the
arguments needed