file_model: Improve sort key performance
Don't query for the actual GFile in the list, because we already have
the information from the iterator. Cache the other path segments since
they are likely to be repeated.

On a local SMB share folder with 10,000 images and 4 path segments, this
reduces the loading time from 30 seconds to 1 second.

Fixes #534


(cherry picked from commit 71351a0c033368f65716087fb557e24356c8eca8)

Co-authored-by: Sophie Herold <sophie@hemio.de>